+++ tests/test_plugins_ua_fallback.py
@@ -1,4 +1,4 @@
 def test_plugin():
-    from reader._plugins.cloudflare_ua_fix import init
+    from reader._plugins.ua_fallback import init
 
     # TODO: actually test the plugin